refactor: move layout components out of layout file
